refactor: modularize app controller
