+++ src/js/app.js
@@ -14,6 +14,8 @@ import { projectManagement } from './projectManagement.js'
 import { modeManagement } from './modeManagement.js'
 import { commandLibrary } from './commandLibrary.js'
 import { aliasView } from './aliasView.js'
+import { viewManagement } from './viewManagement.js'
+import { welcome } from './welcome.js'
 
 export default class STOToolsKeybindManager {
   constructor() {
@@ -162,290 +164,6 @@ export default class STOToolsKeybindManager {
       eventBus.emit('sto-app-error', { error })
     }
   }
-  updateViewToggleButton(viewMode) {
-    const toggleBtn = document.getElementById('toggleKeyViewBtn')
-    if (toggleBtn) {
-      const icon = toggleBtn.querySelector('i')
-      if (viewMode === 'categorized') {
-        icon.className = 'fas fa-sitemap'
-        toggleBtn.title = 'Switch to key type view'
-      } else if (viewMode === 'key-types') {
-        icon.className = 'fas fa-th'
-        toggleBtn.title = 'Switch to grid view'
-      } else {
-        icon.className = 'fas fa-list'
-        toggleBtn.title = 'Switch to command categories'
-      }
-    }
-  }
-
-  toggleKeyView() {
-    // Only allow view mode changes in keybind mode, not alias mode
-    if (this.currentEnvironment === 'alias') {
-      return
-    }
-    
-    const currentMode = localStorage.getItem('keyViewMode') || 'key-types'
-    let newMode
-
-    // 3-way toggle: key-types → grid → categorized → key-types
-    if (currentMode === 'key-types') {
-      newMode = 'grid'
-    } else if (currentMode === 'grid') {
-      newMode = 'categorized'
-    } else {
-      newMode = 'key-types'
-    }
-
-    localStorage.setItem('keyViewMode', newMode)
-    this.renderKeyGrid()
-    this.updateViewToggleButton(newMode)
-  }
-
-  // Utility Methods
-  
-  saveCurrentBuild() {
-    const profile = stoStorage.getProfile(this.currentProfile)
-    const currentBuild = this.getCurrentProfile()
-
-    if (profile && currentBuild) {
-      // Ensure builds structure exists
-      if (!profile.builds) {
-        profile.builds = {
-          space: { keys: {} },
-          ground: { keys: {} },
-        }
-      }
-
-      // Save current build data
-      profile.builds[this.currentEnvironment] = {
-        keys: currentBuild.keys || {},
-        // Note: aliases are profile-level, not build-specific
-      }
-
-      stoStorage.saveProfile(this.currentProfile, profile)
-    }
-  }
-
-  async confirmDeleteKey(keyName) {
-    const confirmed = await stoUI.confirm(
-      i18next.t('confirm_delete_key', { keyName }),
-      i18next.t('delete_key'),
-      'danger'
-    )
-
-    if (confirmed) {
-      this.deleteKey(keyName)
-    }
-  }
-
-  async confirmClearChain(keyName) {
-    const confirmed = await stoUI.confirm(
-      i18next.t('confirm_clear_commands', { keyName }),
-      i18next.t('clear_commands'),
-      'warning'
-    )
-
-    if (confirmed) {
-      const profile = this.getCurrentProfile()
-      profile.keys[keyName] = []
-      this.saveCurrentBuild() // Use saveCurrentBuild for build-level data
-      this.renderCommandChain()
-      this.renderKeyGrid()
-      this.setModified(true)
-
-      stoUI.showToast(i18next.t('commands_cleared_for_key', {keyName: keyName}), 'success')
-    }
-  }
-
-
-  isFirstTime() {
-    return !localStorage.getItem('sto_keybind_manager_visited')
-  }
-
-  showWelcomeMessage() {
-    localStorage.setItem('sto_keybind_manager_visited', 'true')
-    modalManager.show('aboutModal')
-  }
-
-  // Additional Methods
-  duplicateKey(keyName) {
-    const profile = this.getCurrentProfile()
-    const commands = profile.keys[keyName]
-
-    if (!commands || commands.length === 0) {
-      stoUI.showToast(i18next.t('no_commands_to_duplicate'), 'warning')
-      return
-    }
-
-    // Find a suitable new key name
-    let newKeyName = keyName + '_copy'
-    let counter = 1
-
-    while (profile.keys[newKeyName]) {
-      newKeyName = `${keyName}_copy_${counter}`
-      counter++
-    }
-
-    // Clone commands
-    const clonedCommands = commands.map((cmd) => ({
-      ...cmd,
-      id: this.generateCommandId(),
-    }))
-
-    profile.keys[newKeyName] = clonedCommands
-    stoStorage.saveProfile(this.currentProfile, profile)
-    this.renderKeyGrid()
-    this.setModified(true)
-
-    stoUI.showToast(i18next.t('key_duplicated', {keyName: keyName, newKeyName: newKeyName}), 'success')
-  }
-
-  showTemplateModal() {
-    stoUI.showToast(i18next.t('template_system_coming_soon'))
-  }
-
-  validateCurrentChain() {
-    if (!this.selectedKey) {
-      stoUI.showToast(i18next.t('no_key_selected'), 'warning')
-      return
-    }
-
-    const profile = this.getCurrentProfile()
-    const commands = profile.keys[this.selectedKey] || []
-
-    if (commands.length === 0) {
-      stoUI.showToast(i18next.t('no_commands_to_validate'), 'warning')
-      return
-    }
-
-    const validation = stoKeybinds.validateKeybind(this.selectedKey, commands)
-
-    if (validation.valid) {
-      stoUI.showToast(i18next.t('command_chain_is_valid'), 'success')
-    } else {
-      const errorMsg = 'Validation errors:\n' + validation.errors.join('\n')
-      stoUI.showToast(i18next.t('error_message', {error: errorMsg}), 'error', 5000)
-    }
-  }
-
-  filterKeys(filter) {
-    const filterLower = filter.toLowerCase()
-
-    // Filter grid view keys (.key-item)
-    const keyItems = document.querySelectorAll('.key-item')
-    keyItems.forEach((item) => {
-      const keyName = item.dataset.key.toLowerCase()
-      const visible = !filter || keyName.includes(filterLower)
-      item.style.display = visible ? 'flex' : 'none'
-    })
-
-    // Filter categorized/key-type view keys (.command-item[data-key])
-    const commandItems = document.querySelectorAll('.command-item[data-key]')
-    commandItems.forEach((item) => {
-      const keyName = item.dataset.key.toLowerCase()
-      const visible = !filter || keyName.includes(filterLower)
-      item.style.display = visible ? 'flex' : 'none'
-    })
-
-    // Hide/show categories based on whether they have visible keys
-    const categories = document.querySelectorAll('.category')
-    categories.forEach((category) => {
-      const visibleKeys = category.querySelectorAll(
-        '.command-item[data-key]:not([style*="display: none"])'
-      )
-      const categoryVisible = !filter || visibleKeys.length > 0
-      category.style.display = categoryVisible ? 'block' : 'none'
-    })
-  }
-
-  filterCommands(filter) {
-    const commandItems = document.querySelectorAll('.command-item')
-    const filterLower = filter.toLowerCase()
-
-    // Filter command items
-    commandItems.forEach((item) => {
-      const text = item.textContent.toLowerCase()
-      const visible = !filter || text.includes(filterLower)
-      item.style.display = visible ? 'flex' : 'none'
-    })
-
-    // Hide/show categories based on whether they have visible commands
-    const categories = document.querySelectorAll('.category')
-    categories.forEach((category) => {
-      const visibleCommands = category.querySelectorAll(
-        '.command-item:not([style*="display: none"])'
-      )
-      const categoryVisible = !filter || visibleCommands.length > 0
-      category.style.display = categoryVisible ? 'block' : 'none'
-    })
-  }
-
-  showAllKeys() {
-    // Show all grid view keys
-    const keyItems = document.querySelectorAll('.key-item')
-    keyItems.forEach((item) => {
-      item.style.display = 'flex'
-    })
-
-    // Show all categorized/key-type view keys
-    const commandItems = document.querySelectorAll('.command-item[data-key]')
-    commandItems.forEach((item) => {
-      item.style.display = 'flex'
-    })
-
-    // Show all categories
-    const categories = document.querySelectorAll('.category')
-    categories.forEach((category) => {
-      category.style.display = 'block'
-    })
-
-    const filterInput = document.getElementById('keyFilter')
-    if (filterInput) {
-      filterInput.value = ''
-    }
-  }
-
-  toggleLibrary() {
-    const content = document.getElementById('libraryContent')
-    const btn = document.getElementById('toggleLibraryBtn')
-
-    if (content && btn) {
-      const isCollapsed = content.style.display === 'none'
-      content.style.display = isCollapsed ? 'block' : 'none'
-
-      const icon = btn.querySelector('i')
-      if (icon) {
-        icon.className = isCollapsed
-          ? 'fas fa-chevron-up'
-          : 'fas fa-chevron-down'
-      }
-    }
-  }
-
-  saveCommandFromModal() {
-    if (!this.selectedKey) {
-      stoUI.showToast(i18next.t('please_select_a_key_first'), 'warning')
-      return
-    }
-
-    const command = stoCommands.getCurrentCommand()
-    if (!command) {
-      stoUI.showToast(i18next.t('please_configure_a_command'), 'warning')
-      return
-    }
-
-    const validation = stoCommands.validateCommand(command)
-    if (!validation.valid) {
-      stoUI.showToast(i18next.t('validation_error'), 'error')
-      return
-    }
-
-    this.addCommand(this.selectedKey, command)
-    modalManager.hide('addCommandModal')
-  }
-
-  // Utility Methods
 }
 
 // Initialize application
@@ -461,6 +179,8 @@ Object.assign(
   modeManagement,
   commandLibrary,
   aliasView,
+  viewManagement,
+  welcome,
 )
 ;
